Rename back to register package
