+++ app/src/main/res/navigation/nav_graph.xml
@@ -14,7 +14,12 @@
 
     <fragment
         android:id="@+id/home_fragment"
-        android:name="com.example.home.HomeFragment" />
+        android:name="com.example.home.HomeFragment">
+        <argument
+            android:name="helloText"
+            android:defaultValue="blah"
+            app:argType="string" />
+    </fragment>
 
     <include app:graph="@navigation/user_nav_graph" />
