+++ user/src/main/java/com/example/user/presentation/register/RegisterMviView.kt
@@ -1,16 +1,16 @@
-package com.example.user.presentation.register2
+package com.example.user.presentation.register
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.example.base.presentation.BaseMviView
-import com.example.user.databinding.FragmentRegister2Binding
+import com.example.user.databinding.FragmentRegisterBinding
 
 class RegisterMviView(
         layoutInflater: LayoutInflater,
         parent: ViewGroup?,
         override val acceptIntent: (RegisterIntent) -> Unit
 ) : BaseMviView<RegisterViewState, RegisterIntent>() {
-    private val binding = FragmentRegister2Binding.inflate(layoutInflater, parent, false)
+    private val binding = FragmentRegisterBinding.inflate(layoutInflater, parent, false)
     override val rootView = binding.root
 
     init {
