+++ user/src/main/java/com/example/user/presentation/register/RegisterPresenter.kt
@@ -1,4 +1,4 @@
-package com.example.user.presentation.register2
+package com.example.user.presentation.register
 
 import androidx.hilt.lifecycle.ViewModelInject
 import com.example.base.presentation.BasePresenter
