+++ user/src/main/java/com/example/user/presentation/register/RegisterViewEvent.kt
@@ -1,4 +1,4 @@
-package com.example.user.presentation.register2
+package com.example.user.presentation.register
 
 sealed class RegisterViewEvent {
     object GoBack: RegisterViewEvent()
