+++ user/src/main/java/com/example/user/presentation/register/RegisterViewState.kt
@@ -1,4 +1,4 @@
-package com.example.user.presentation.register2
+package com.example.user.presentation.register
 
 import java.io.Serializable
