+++ user/src/main/res/layout/fragment_register.xml
