+++ user/src/main/res/navigation/user_nav_graph.xml
@@ -11,7 +11,11 @@
         tools:layout="@layout/fragment_login">
         <action
             android:id="@+id/action_login_to_home"
-            app:destination="@+id/home_fragment" />
+            app:destination="@+id/home_fragment">
+            <argument
+                android:name="helloText"
+                app:argType="string" />
+        </action>
         <action
             android:id="@+id/action_login_fragment_to_register_fragment"
             app:destination="@id/register_fragment" />
@@ -19,7 +23,7 @@
 
     <fragment
         android:id="@+id/register_fragment"
-        android:name="com.example.user.presentation.register2.RegisterFragment2"
-        tools:layout="@layout/fragment_register2" />
+        android:name="com.example.user.presentation.register.RegisterFragment"
+        tools:layout="@layout/fragment_register" />
 
 </navigation>
